Expression: Move algebraic and inline operations to operations.py
We still import them to symbols so that `sym.Sum` works as before.
